pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rio-virtualized-list"
version = "0.2.0"
description = "High-performance virtualized list component for Rio"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "rio-ui>=0.12",
]

[project.urls]
Repository = "https://github.com/JGroxz/rio-component-virtualized-list"

[tool.hatch.build.targets.wheel]
packages = ["src/rio_virtualized_list"]
